mingw: Fix unlink for open files
Files that were opened by a (possibly another) process either for
read or for failed to be replaced by Git on checkout.

If a file is opened with FILE_SHARE_DELETE share mode, it is possible
to delete or rename the file, but it is *not possible* to create a
new file until the file's handle is closed.

To overcome this, unlink now renames files that are not writable before
actually unlinking them.

If rename fails, then the file either doesn't exist, or it is open
without share permissions. On this case, it is still possible to unlink
the file. The file will effectively be deleted when it is closed.
On this case, preserve the existing behavior.

If rename succeeds, unlink the temporary file, making it possible for
the real file name to be reused.

Fixes #1653.

Signed-off-by: Orgad Shaneh <orgads@gmail.com>